mediapipeの導入完了。コンソールで結果出力を確認。
