+++ front/src/app/components/BasicSlider/swiper.module.css
