+++ front/src/app/page.tsx
@@ -1,9 +1,9 @@
 
 "use client";
 
-import BasicSlider from "./components/BasicSlider";
+import BasicSlider from "./components/BasicSlider/Swiper";
 import styles from "./page.module.css";
-import { Mediapipe } from "./Recording/Mediapipe";
+import { Mediapipe } from "./components/Detection/Mediapipe";
 
 export default function Home() {
   return (
